requirements.txt
streamlit>=0.55.0
Pillow>=8.0.0
opencv-python>=4.5.0
numpy>=1.19.0
pandas>=1.1.0
requests>=2.24.0
huggingface-hub>=0.8.0 